feat: rename llm_models
